+++ db/migrate/20150326184844_add_progress_report_data_fields_to_applicants.rb
@@ -1,0 +1,8 @@
+class AddProgressReportDataFieldsToApplicants < ActiveRecord::Migration
+  def change
+    add_column :applicants, :approved_irb_protocols, :text
+    add_column :applicants, :approved_irb_document, :string
+    add_column :applicants, :approved_iacuc_protocols, :text
+    add_column :applicants, :approved_iacuc_document, :string
+  end
+end